pyproject.toml
[tool.poetry]
name = "spt-webui-backend"
version = "0.1.0"
description = ""
authors = ["Niklas Dietzel <niklas@dietzel.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pydantic-settings = "^2.3.4"
oauthlib = "^3.2.2"
requests-oauthlib = "^2.0.0"
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"
alembic = "^1.13.2"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
